Adjusts frequency of database downloads
We were downloading a ton of data during csv imports because we were downloading the entire database after each upload. Now it only downloads the database once at the end of the csv import. We lose the cool view of watching the movies populate the header grid but we'll just have to live without that.